Move wire app initializer into "internal/app" package
